Move definitions for template SLB_cin_char.bt to a single file
